@@ -1,15 +1,9 @@
 package QuickNotes;
 
-import android.annotation.SuppressLint;
-import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import androidx.appcompat.app.ActionBar;
-import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -21,15 +15,20 @@ import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
+import java.util.Locale;
 
 public class NotesInFolderPage extends AppCompatActivity {
     TextView folderNameView;
-    static ListView notesFolderListView;
+    ListView notesFolderListView;
     Button newNotesBtn;
     String currentFolder;
     static ArrayList<String> allNotesContent = new ArrayList<>();
@@ -37,10 +36,11 @@ public class NotesInFolderPage extends AppCompatActivity {
     static ArrayList<String> allNoteDates = new ArrayList<>();
     String filePath;
     Date dateCreated;
-    static NotesCustomAdapter notesAdapter;
+    NotesCustomAdapter notesAdapter;
     SharedPreferences pref;
     private Boolean nightTheme;
     Context context;
+    Boolean noteIsClicked;
 
     protected void onCreate(Bundle savedInstanceState) {
         pref = getSharedPreferences("MyPref", 0);
@@ -50,7 +50,6 @@ public class NotesInFolderPage extends AppCompatActivity {
         else if (! pref.getBoolean("NIGHT MODE", false)) {
             setTheme(R.style.AppTheme);
         }
-
         nightTheme = pref.getBoolean("NIGHT MODE", false);
         setContentView(R.layout.notes_in_folder_view);
         Toolbar toolbar = findViewById(R.id.toolbar);
@@ -60,7 +59,7 @@ public class NotesInFolderPage extends AppCompatActivity {
         actionBar.setHomeButtonEnabled(true);
         actionBar.setDisplayHomeAsUpEnabled(true);
         super.onCreate(savedInstanceState);
-        folderNameView = findViewById(R.id.folderNameView);
+        folderNameView = findViewById(R.id.noteNameView);
         notesFolderListView = findViewById(R.id.notesFolderListView);
         newNotesBtn = findViewById(R.id.newNoteBtn);
         filePath = getFilesDir().getAbsolutePath();
@@ -73,98 +72,79 @@ public class NotesInFolderPage extends AppCompatActivity {
         notesFolderListView.setAdapter(notesAdapter);
         context = getBaseContext();
 
-        notesFolderListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                Intent intent = new Intent(view.getContext(), SpecificNoteEdit.class);
-                String clickedNoteName = notesFolderListView.getItemAtPosition(position).toString();
-                String clickedNoteContent = allNotesContent.get(position);
-                String clickedNoteDate = allNoteDates.get(position);
-                intent.putExtra("Folder name", currentFolder);
-                intent.putExtra("Note name", clickedNoteName);
-                intent.putExtra("Note content", clickedNoteContent);
-                intent.putExtra("Note date", clickedNoteDate);
-                startActivity(intent);
-            }
+        notesFolderListView.setOnItemClickListener((AdapterView<?> parent, View view, int position, long id) -> {
+                // Check if a folder intent has not been clicked yet in the listview.
+                if (!noteIsClicked) {
+                    noteIsClicked = true;
+                    Intent intent = new Intent(view.getContext(), SpecificNoteEdit.class);
+                    String clickedNoteName = notesFolderListView.getItemAtPosition(position).toString();
+                    String clickedNoteContent = allNotesContent.get(position);
+                    String clickedNoteDate = allNoteDates.get(position);
+                    intent.putExtra("Folder name", currentFolder);
+                    intent.putExtra("Note name", clickedNoteName);
+                    intent.putExtra("Note content", clickedNoteContent);
+                    intent.putExtra("Note date", clickedNoteDate);
+                    startActivity(intent);
+                }
         });
 
-        newNotesBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(final View view) {
+        newNotesBtn.setOnClickListener((View view) -> {
                 AlertDialog.Builder builder = new AlertDialog.Builder(view.getContext());
                 View alertView = View.inflate(context, R.layout.alertdialog_new_note, null);
                 final Button positiveButton = alertView.findViewById(R.id.positiveButton);
                 final Button cancelButton = alertView.findViewById(R.id.cancelButton);
                 final EditText noteNameText = alertView.findViewById(R.id.noteNameText);
+                noteNameText.requestFocus();
+                // Force a keyboard for the edit text.
+                InputMethodManager inputMethodManager = (InputMethodManager) context.getSystemService(Context.INPUT_METHOD_SERVICE);
+                inputMethodManager.toggleSoftInput(InputMethodManager.SHOW_FORCED, 0);
                 final EditText noteContentText = alertView.findViewById(R.id.reminderNameText);
                 builder.setView(alertView);
                 final AlertDialog optionDialog = builder.show();
-                positiveButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        final String noteNameString = noteNameText.getText().toString();
-                        final String noteContentString = noteContentText.getText().toString();
-                        if (noteNameString.equals("")) {
-                            noteNameText.setError("Please enter a name.");
-                        }
-                        else if (allNoteNames.contains(noteNameString)){
-                            noteNameText.setError("This note already exists.");
-                        }
-                        else {
-                            dateCreated = Calendar.getInstance().getTime();
-                            SimpleDateFormat formatter = new SimpleDateFormat("MM/dd/yyyy");
-                            String dateString = formatter.format(dateCreated);
-                            Note.saveNote(context, currentFolder, noteNameString, noteContentString, dateString);
-                            allNoteNames.add(0, noteNameString);
-                            allNotesContent.add(0, noteContentString);
-                            allNoteDates.add(0, dateString);
-                            notesAdapter.notifyDataSetChanged();
-                            optionDialog.dismiss();
-                            Toast.makeText(context, "Note saved.", Toast.LENGTH_LONG).show();
-                        }
+                positiveButton.setOnClickListener(view12 -> {
+                    final String noteNameString = noteNameText.getText().toString();
+                    final String noteContentString = noteContentText.getText().toString();
+                    if (noteNameString.equals("")) {
+                        noteNameText.setError("Please enter a name.");
+                    }
+                    else if (allNoteNames.contains(noteNameString)){
+                        noteNameText.setError("This note already exists.");
+                    }
+                    else {
+                        dateCreated = Calendar.getInstance().getTime();
+                        SimpleDateFormat formatter = new SimpleDateFormat("MM/dd/yyyy", Locale.US);
+                        String dateString = formatter.format(dateCreated);
+                        Note.saveNote(context, currentFolder, noteNameString, noteContentString, dateString);
+                        allNoteNames.add(0, noteNameString);
+                        allNotesContent.add(0, noteContentString);
+                        allNoteDates.add(0, dateString);
+                        notesAdapter.notifyDataSetChanged();
+                        optionDialog.dismiss();
+                        Toast.makeText(context, "Note saved.", Toast.LENGTH_LONG).show();
                     }
                 });
-                cancelButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        optionDialog.dismiss();
-
-                }
-            });
-            }
+                cancelButton.setOnClickListener(view1 -> optionDialog.dismiss());
         });
 
-        notesFolderListView.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
-            @SuppressLint("SetTextI18n")
-            @Override
-            public boolean onItemLongClick(AdapterView<?> parent, View view, final int position, long id) {
-                final String noteName = notesFolderListView.getItemAtPosition(position).toString();
-                AlertDialog.Builder builder = new AlertDialog.Builder(NotesInFolderPage.this);
-                View alertView = View.inflate(context, R.layout.alertdialog_delete_note, null);
-                final Button positiveButton = alertView.findViewById(R.id.positiveButton);
-                final Button cancelButton = alertView.findViewById(R.id.cancelButton);
-                builder.setView(alertView);
-                final AlertDialog optionDialog = builder.show();
-                positiveButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        Note.deleteNote(context, noteName, currentFolder);
-                        allNoteNames.remove(noteName);
-                        allNotesContent.remove(Note.loadNote(context, noteName, currentFolder));
-                        allNoteDates.remove(Note.loadDate(context, noteName, currentFolder));
-                        notesAdapter.notifyDataSetChanged();
-                        Toast.makeText(context, "Note deleted.", Toast.LENGTH_LONG).show();
-                        optionDialog.dismiss();
-                    }
-                });
-                cancelButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        optionDialog.dismiss();
-                    }
-                });
-                return true;
-            }
+        notesFolderListView.setOnItemLongClickListener((parent, view, position, id) -> {
+            final String noteName = notesFolderListView.getItemAtPosition(position).toString();
+            AlertDialog.Builder builder = new AlertDialog.Builder(NotesInFolderPage.this);
+            View alertView = View.inflate(context, R.layout.alertdialog_delete_note, null);
+            final Button positiveButton = alertView.findViewById(R.id.positiveButton);
+            final Button cancelButton = alertView.findViewById(R.id.cancelButton);
+            builder.setView(alertView);
+            final AlertDialog optionDialog = builder.show();
+            positiveButton.setOnClickListener(view14 -> {
+                Note.deleteNote(context, noteName, currentFolder);
+                allNoteNames.remove(noteName);
+                allNotesContent.remove(Note.loadNote(context, noteName, currentFolder));
+                allNoteDates.remove(Note.loadDate(context, noteName, currentFolder));
+                notesAdapter.notifyDataSetChanged();
+                Toast.makeText(context, "Note deleted.", Toast.LENGTH_LONG).show();
+                optionDialog.dismiss();
+            });
+            cancelButton.setOnClickListener(view13 -> optionDialog.dismiss());
+            return true;
         });
 
     }
@@ -175,20 +155,12 @@ public class NotesInFolderPage extends AppCompatActivity {
         return true;
     }
 
-    public void hideKeyboard(Activity activity) {
-        View view = activity.getCurrentFocus();
-        if (view != null) {
-            InputMethodManager imm = (InputMethodManager) activity.getSystemService(Context.INPUT_METHOD_SERVICE);
-            imm.hideSoftInputFromWindow(view.getWindowToken(), 0);
-        }
-    }
-
-    public void refreshListView(Context context, String currentFolder) {
+    public void refreshListView(Context context, String currentFolder, ListView listView) {
             allNoteNames = Note.loadAllNoteNamesInFolder(context, currentFolder);
             allNotesContent = Note.loadAllNoteContentInFolder(context, currentFolder);
             allNoteDates = Note.loadAllNoteDateInFolder(context, currentFolder);
             notesAdapter = new NotesCustomAdapter(context, allNoteNames, allNotesContent, allNoteDates);
-            notesFolderListView.setAdapter(notesAdapter);
+            listView.setAdapter(notesAdapter);
     }
 
     @Override
@@ -222,7 +194,8 @@ public class NotesInFolderPage extends AppCompatActivity {
         if (nightTheme != pref.getBoolean("NIGHT MODE", false)){
             recreate();
         }
+        noteIsClicked = false;
+        refreshListView(context, currentFolder, notesFolderListView);
         super.onResume();
     }
-
 }
