@@ -1,6 +1,5 @@
 package QuickNotes;
 
-import android.annotation.SuppressLint;
 import android.content.Context;
 import android.util.Log;
 
@@ -13,17 +12,14 @@ import java.io.FileOutputStream;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
-import java.util.Date;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Scanner;
 
 class Note {
-    @SuppressLint("NewApi")
     static void saveNote(Context context, String folderName, String noteName, String content, String dateString) {
         String filePath = context.getFilesDir().getAbsolutePath() + "/" + "notes_folders";
         try {
@@ -47,7 +43,8 @@ class Note {
     static void saveFolderName(Context context, String folderName) {
         try {
             File folder = new File(context.getFilesDir() + "/notes_folders", folderName);
-            folder.mkdir();
+            Boolean result = folder.mkdir();
+            Log.d("Create result: ", result + "");
             // append to file
             FileOutputStream fos = context.openFileOutput("folder_names.txt", Context.MODE_APPEND);
             fos.write(folderName.getBytes());
@@ -58,14 +55,14 @@ class Note {
         }
     }
 
-    private static boolean fileExists(Context context, String fileName) {
+    private static boolean fileExists(Context context) {
         String filePath = context.getFilesDir().getAbsolutePath();
-        File file = new File(filePath, fileName);
+        File file = new File(filePath, "folder_names.txt");
         return file.exists();
     }
 
-    private static boolean folderExists(Context context, String folderName) {
-        File file = new File(context.getFilesDir().getAbsolutePath() + "/" + "notes_folders" + "/" + folderName);
+    private static boolean folderExists(Context context) {
+        File file = new File(context.getFilesDir().getAbsolutePath() + "/" + "notes_folders" + "/" + "notes_folders");
         return file.isDirectory();
     }
 
@@ -125,7 +122,7 @@ class Note {
         StringBuilder noteContent = new StringBuilder();
         int counter = 0;
         try {
-            FileInputStream inputStream = new FileInputStream(new File(filePath));
+            FileInputStream inputStream = new FileInputStream(filePath);
             InputStreamReader inputStreamReader = new InputStreamReader(inputStream);
             BufferedReader bufferedReader = new BufferedReader(inputStreamReader);
             String receiveString;
@@ -152,7 +149,7 @@ class Note {
         String filePath = context.getFilesDir() + "/" + "notes_folders" + "/" + folderName + "/" + noteName;
         StringBuilder noteDate = new StringBuilder();
         try {
-            FileInputStream inputStream = new FileInputStream(new File(filePath));
+            FileInputStream inputStream = new FileInputStream(filePath);
             InputStreamReader inputStreamReader = new InputStreamReader(inputStream);
             BufferedReader bufferedReader = new BufferedReader(inputStreamReader);
             String receiveString;
@@ -171,15 +168,17 @@ class Note {
         ArrayList<String> folderNamesList = new ArrayList<>();
         String filePath = context.getFilesDir().getAbsolutePath();
         try {
-            if (!folderExists(context, "notes_folders")) {
+            if (!folderExists(context)) {
                 File folder = new File(filePath, "notes_folders");
-                folder.mkdir();
+                Boolean result1 = folder.mkdir();
+                Log.d("Create result1: ", result1 + "");
                 File folder1 = new File(filePath + "/" + "notes_folders", "Unclassified");
-                folder1.mkdir();
+                Boolean result2 = folder1.mkdir();
+                Log.d("Create result2: ", result2 + "");
             }
 
-            if (fileExists(context, "folder_names.txt")) {
-                FileInputStream inputStream = new FileInputStream(new File(filePath + "/folder_names.txt"));
+            if (fileExists(context)) {
+                FileInputStream inputStream = new FileInputStream(filePath + "/folder_names.txt");
                 InputStreamReader inputStreamReader = new InputStreamReader(inputStream);
                 BufferedReader bufferedReader = new BufferedReader(inputStreamReader);
                 String receiveString;
@@ -190,7 +189,6 @@ class Note {
             } else {
                 File file = new File(filePath, "folder_names.txt");
                 FileOutputStream fos = new FileOutputStream(file, true);
-                //FileOutputStream fos = context.openFileOutput(filePath + "/" + "notes_folders" + "/" + "folder_names.txt", Context.MODE_APPEND);
                 folderNamesList.add("Unclassified");
                 fos.write("Unclassified".getBytes());
                 fos.write('\n');
@@ -214,18 +212,20 @@ class Note {
     static void deleteNote( Context context, String noteName, String folderName) {
         String filePath = context.getFilesDir().getAbsolutePath() + "/notes_folders" + "/" + folderName + "/" + noteName + ".txt";
         File fileOrDirectory = new File(filePath);
-        fileOrDirectory.delete();
+        Boolean result = fileOrDirectory.delete();
+        Log.d("Delete result: ", result + "");
     }
 
     static void deleteFolder(Context context, String folderName) {
         String filePath = context.getFilesDir().getAbsolutePath() + "/notes_folders/" + folderName;
         File fileOrDirectory = new File(filePath);
         if (fileOrDirectory.isDirectory()) {
-            for (File child : fileOrDirectory.listFiles()) {
+            for (File child : Objects.requireNonNull(fileOrDirectory.listFiles())) {
                 deleteFolder(context, folderName + "/" + child.getName());
             }
         }
-        fileOrDirectory.delete();
+        Boolean result = fileOrDirectory.delete();
+        Log.d("Delete result: ", result + "");
         try {
             filePath = context.getFilesDir().getAbsolutePath();
             File file = new File(filePath, "folder_names.txt");
@@ -235,7 +235,6 @@ class Note {
         }
     }
 
-    @SuppressLint("Assert")
     private static void removeLine(final File file, final String line) throws IOException {
         final List<String> lines = new LinkedList<>();
         final Scanner reader = new Scanner(new FileInputStream(file), "UTF-8");
