@@ -10,7 +10,11 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.graphics.Color;
 import android.os.Build;
+import android.util.Log;
+
 import androidx.core.app.NotificationCompat;
+
+import java.util.Objects;
 
 
 // idk what im doing because i want to make this have specific title and content for this to show as a notification.
@@ -18,7 +22,6 @@ public class ReminderBroadcast extends BroadcastReceiver {
     SharedPreferences pref;
     NotificationManager notificationManager;
     int notificationIDCounter;
-    SharedPreferences.Editor editor;
     String noteName;
     String noteContent;
     String currentFolder;
@@ -26,19 +29,20 @@ public class ReminderBroadcast extends BroadcastReceiver {
 
     @Override
     public void onReceive(Context context, Intent intent) {
+        Log.d("Intent action: ", "" + intent.getAction());
         pref = context.getSharedPreferences("MyPref", 0);
         notificationIDCounter = pref.getInt("notificationID", 0);
-        final int NOTIFY_ID =  notificationIDCounter - 1;// ID of notification
+        final int NOTIFY_ID = notificationIDCounter - 1;// ID of notification
         String id = "Note Reminder"; // default_channel_id
         String title = "title"; // Default Channel
         currentFolder = pref.getString("Note folder", "Default");
         noteName = pref.getString("Note name", "Default");
         noteDate = pref.getString("Note date", "Default");
         noteContent = pref.getString("Note content", "Default");
-        PendingIntent pendingIntent;
+        PendingIntent pendingIntent = null;
         NotificationCompat.Builder builder;
         if (notificationManager == null) {
-            notificationManager = (NotificationManager)context.getSystemService(Context.NOTIFICATION_SERVICE);
+            notificationManager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
         }
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             int importance = NotificationManager.IMPORTANCE_HIGH;
@@ -53,7 +57,7 @@ public class ReminderBroadcast extends BroadcastReceiver {
             intent = new Intent(context, HomePage.class);
 
             intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_SINGLE_TOP);
-            pendingIntent = PendingIntent.getActivity(context, 0, intent, 0);
+            pendingIntent = PendingIntent.getActivity(context, 0, intent, PendingIntent.FLAG_IMMUTABLE);
             builder.setContentTitle("Reminder")
                     .setSmallIcon(R.drawable.outline_notifications_black)
                     .setContentText(noteName)
@@ -65,13 +69,14 @@ public class ReminderBroadcast extends BroadcastReceiver {
                     .setLights(Color.BLUE, 3000, 3000)
                     .setStyle(new NotificationCompat.BigTextStyle()
                             .bigText(noteContent));
-        }
-        else {
+        } else {
             builder = new NotificationCompat.Builder(context, id);
             intent = new Intent(context, HomePage.class);
 
             intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_SINGLE_TOP);
-            pendingIntent = PendingIntent.getActivity(context, 0, intent, 0);
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                pendingIntent = PendingIntent.getActivity(context, 0, intent, PendingIntent.FLAG_IMMUTABLE);
+            }
             builder.setContentTitle("Reminder")
                     .setSmallIcon(R.drawable.outline_notifications_black)
                     .setContentText(noteName)
