@@ -1,10 +1,10 @@
 package QuickNotes;
 
-import android.annotation.SuppressLint;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.os.Build;
 import android.os.Bundle;
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AlertDialog;
@@ -12,7 +12,6 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
 import android.view.View;
-import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.ListView;
 import android.widget.TextView;
@@ -21,7 +20,7 @@ import android.widget.Toast;
 import java.util.ArrayList;
 
 public class RemindersPage extends AppCompatActivity {
-    static ListView reminderListView;
+    ListView reminderListView;
     Button newNotesBtn;
     static ArrayList<String> allReminderNames = new ArrayList<>();
     static ArrayList<String> allNotesContent = new ArrayList<>();
@@ -31,7 +30,7 @@ public class RemindersPage extends AppCompatActivity {
     String reminderNotificationID;
     Intent intent;
     PendingIntent pendingIntent;
-    static RemindersCustomAdapter notesAdapter;
+    RemindersCustomAdapter notesAdapter;
     SharedPreferences pref;
     private Boolean nightTheme;
 
@@ -55,55 +54,47 @@ public class RemindersPage extends AppCompatActivity {
         newNotesBtn = findViewById(R.id.newNoteBtn);
         allReminderNames = Reminder.loadReminderNames(this);
         allNotesContent = Reminder.loadAllNoteContentInReminderFolder(this);
-        notesAdapter = new RemindersCustomAdapter(this, allReminderNames, allNotesContent);
+        notesAdapter = new RemindersCustomAdapter(this, allReminderNames);
         reminderListView.setAdapter(notesAdapter);
         context = getBaseContext();
 
 
-        reminderListView.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
-            @SuppressLint("SetTextI18n")
-            @Override
-            public boolean onItemLongClick(AdapterView<?> parent, View view, final int position, long id) {
-                final String longClickedNoteName = reminderListView.getItemAtPosition(position).toString();
-                AlertDialog.Builder builder = new AlertDialog.Builder(RemindersPage.this);
-                View alertView = View.inflate(context, R.layout.alertdialog_delete_reminder, null);
-                final Button positiveButton = alertView.findViewById(R.id.positiveButton);
-                final Button cancelButton = alertView.findViewById(R.id.cancelButton);
-                builder.setView(alertView);
-                final AlertDialog optionDialog = builder.show();
-                reminderNameText = findViewById(R.id.noteNameText);
-                reminderDateText = findViewById(R.id.reminderDateText);
-                StringBuilder usingJustToReverse = new StringBuilder(longClickedNoteName);
-                String reversedLongClickedName = usingJustToReverse.reverse().toString();
-                int startNotificationID = longClickedNoteName.length() + 1;
-                for (int i = 0; i < reversedLongClickedName.length(); i++){
-                    startNotificationID--;
-                    if (reversedLongClickedName.charAt(i) == '_'){
-                        break;
-                    }
+        reminderListView.setOnItemLongClickListener((parent, view, position, id) -> {
+            final String longClickedNoteName = reminderListView.getItemAtPosition(position).toString();
+            AlertDialog.Builder builder = new AlertDialog.Builder(RemindersPage.this);
+            View alertView = View.inflate(context, R.layout.alertdialog_delete_reminder, null);
+            final Button positiveButton = alertView.findViewById(R.id.positiveButton);
+            final Button cancelButton = alertView.findViewById(R.id.cancelButton);
+            builder.setView(alertView);
+            final AlertDialog optionDialog = builder.show();
+            reminderNameText = findViewById(R.id.noteNameText);
+            reminderDateText = findViewById(R.id.reminderDateText);
+            StringBuilder usingJustToReverse = new StringBuilder(longClickedNoteName);
+            String reversedLongClickedName = usingJustToReverse.reverse().toString();
+            int startNotificationID = longClickedNoteName.length() + 1;
+            for (int i = 0; i < reversedLongClickedName.length(); i++){
+                startNotificationID--;
+                if (reversedLongClickedName.charAt(i) == '_'){
+                    break;
                 }
-                reminderNotificationID = longClickedNoteName.substring(startNotificationID);
-                positiveButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        Reminder.deleteReminder(context, longClickedNoteName, "");
-                        intent = new Intent(context, ReminderBroadcast.class);
-                        pendingIntent = PendingIntent.getActivity(context, Integer.parseInt(reminderNotificationID), intent, 0);
-                        new SpecificNoteEdit().cancelReminder(context, reminderNotificationID);
-                        refreshReminderListView(context);
-                        reminderListView.setAdapter(notesAdapter);
-                        Toast.makeText(getBaseContext(), "Reminder deleted.", Toast.LENGTH_LONG).show();
-                        optionDialog.dismiss();
-                    }
-                });
-                cancelButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        optionDialog.dismiss();
-                    }
-                });
-                return true;
             }
+            reminderNotificationID = longClickedNoteName.substring(startNotificationID);
+            positiveButton.setOnClickListener(view12 -> {
+                Reminder.deleteReminder(context, longClickedNoteName, "");
+                intent = new Intent(context, ReminderBroadcast.class);
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                    pendingIntent = PendingIntent.getActivity(context, Integer.parseInt(reminderNotificationID), intent, PendingIntent.FLAG_IMMUTABLE);
+                }
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                    new SpecificNoteEdit().cancelReminder(context, reminderNotificationID);
+                }
+                refreshReminderListView(context);
+                reminderListView.setAdapter(notesAdapter);
+                Toast.makeText(getBaseContext(), "Reminder deleted.", Toast.LENGTH_LONG).show();
+                optionDialog.dismiss();
+            });
+            cancelButton.setOnClickListener(view1 -> optionDialog.dismiss());
+            return true;
         });
 
     }
@@ -112,7 +103,7 @@ public class RemindersPage extends AppCompatActivity {
         if (reminderListView != null) {
         allReminderNames = Reminder.loadReminderNames(context);
         allNotesContent = Reminder.loadAllNoteContentInReminderFolder(context);
-        notesAdapter = new RemindersCustomAdapter(context, allReminderNames, allNotesContent);
+        notesAdapter = new RemindersCustomAdapter(context, allReminderNames);
         reminderListView = findViewById(R.id.noteReminderListView);
         reminderListView.setAdapter(notesAdapter);
         }
